+++ src/main/java/ro/brage/dodo/rs/mappers/qualifiers/LoadEntity.java
@@ -16,7 +16,7 @@
  * DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  *******************************************************************************/
-package ro.brage.dodo.jpa.mapper.qualifiers;
+package ro.brage.dodo.rs.mappers.qualifiers;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
